Re-factored mulit-profile support to all exporters
NOTE: Some do not do anything with it by default

The GNU ARM Ecilpse exporter pioneered this multi-profile export support
as part of it's pull request. This commit moves support for multiple
bulid profiles from that one exporter to the Exporter base class. This
allows make, for example, to take advantage of multi-profile exports.